requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
opencv-python
ultralytics
PyYAML
requests
pymodbus
pyserial